+++ scm-ui/ui-webapp/src/repos/sources/components/content/overflowMenu/ContentActionMenu.tsx
@@ -24,14 +24,11 @@
 
 import { binder, extensionPoints } from "@scm-manager/ui-extensions";
 import React, { FC, ReactElement, useState } from "react";
-import { Button, Icon } from "@scm-manager/ui-components";
+import { Icon } from "@scm-manager/ui-components";
 import styled from "styled-components";
 import { Menu } from "@headlessui/react";
-import { useTranslation } from "react-i18next";
-import ActionMenuItem from "./ActionMenuItem";
-import LinkMenuItem from "./LinkMenuItem";
-import ModalMenuItem from "./ModalMenuItem";
-import { Link } from "react-router-dom";
+import FallbackMenuButton from "./FallbackMenuButton";
+import MenuItem from "./MenuItem";
 
 const MenuButton = styled(Menu.Button)`
   background: transparent;
@@ -40,25 +37,6 @@ const MenuButton = styled(Menu.Button)`
   height: 2.5rem;
   width: 50px;
   margin-bottom: 0.5rem;
-`;
-
-const FallbackButton = styled(Button)`
-  height: 2.5rem;
-  width: 50px;
-  margin-bottom: 0.5rem;
-  > i {
-    padding: 0 !important;
-  }
-  &:hover {
-    color: var(--scm-link-color);
-  }
-`;
-
-const FallbackLink = styled(Link)`
-  width: 50px;
-  &:hover {
-    color: var(--scm-link-color);
-  }
 `;
 
 const MenuItems = styled(Menu.Items)`
@@ -86,58 +64,7 @@ type Props = {
   extensionProps: extensionPoints.ContentActionExtensionProps;
 };
 
-const MenuItem: FC<
-  extensionPoints.FileViewActionBarOverflowMenu["type"] &
-    Props & {
-      active: boolean;
-      onClick: (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => void;
-      setSelectedModal: (element: ReactElement | undefined) => void;
-    }
-> = ({ extensionProps, label, icon, props, category, active, onClick, setSelectedModal, ...rest }) => {
-  if ("action" in rest) {
-    return (
-      <ActionMenuItem
-        label={label}
-        icon={icon}
-        category={category}
-        extensionProps={extensionProps}
-        active={active}
-        onClick={onClick}
-        {...rest}
-      />
-    );
-  }
-  if ("link" in rest) {
-    return (
-      <LinkMenuItem
-        category={category}
-        label={label}
-        icon={icon}
-        active={active}
-        extensionProps={extensionProps}
-        {...rest}
-      />
-    );
-  }
-  if ("modalElement" in rest) {
-    return (
-      <ModalMenuItem
-        category={category}
-        label={label}
-        icon={icon}
-        extensionProps={extensionProps}
-        active={active}
-        onClick={onClick}
-        setSelectedModal={setSelectedModal}
-        {...rest}
-      />
-    );
-  }
-  return null;
-};
-
 const ContentActionMenu: FC<Props> = ({ extensionProps }) => {
-  const [t] = useTranslation("plugins");
   const [selectedModal, setSelectedModal] = useState<ReactElement | undefined>();
   const extensions = binder.getExtensions<extensionPoints.FileViewActionBarOverflowMenu>(
     "repos.sources.content.actionbar.menu",
@@ -153,42 +80,6 @@ const ContentActionMenu: FC<Props> = ({ extensionProps }) => {
     },
     {}
   );
-
-  const renderSingleButton = (extension: extensionPoints.FileViewActionBarOverflowMenu["type"]) => {
-    if ("action" in extension) {
-      return (
-        <FallbackButton
-          icon={extension.icon}
-          title={t(extension.label)}
-          action={() => extension.action(extensionProps)}
-        />
-      );
-    }
-    if ("link" in extension) {
-      return (
-        <FallbackLink to={extension.link(extensionProps)} className="button" title={t(extension.label)}>
-          <Icon name={extension.icon} color="inherit" />
-        </FallbackLink>
-      );
-    }
-    if ("modalElement" in extension) {
-      return (
-        <FallbackButton
-          icon={extension.icon}
-          title={t(extension.label)}
-          action={() =>
-            setSelectedModal(
-              React.createElement(extension.modalElement, {
-                ...extensionProps,
-                close: () => setSelectedModal(undefined),
-              })
-            )
-          }
-        />
-      );
-    }
-    return null;
-  };
 
   const renderMenu = () => (
     <>
@@ -237,7 +128,16 @@ const ContentActionMenu: FC<Props> = ({ extensionProps }) => {
 
   return (
     <>
-      {extensions.length === 1 ? renderSingleButton(extensions[0]) : renderMenu()} {selectedModal || null}
+      {extensions.length === 1 ? (
+        <FallbackMenuButton
+          extension={extensions[0]}
+          extensionProps={extensionProps}
+          setSelectedModal={setSelectedModal}
+        />
+      ) : (
+        renderMenu()
+      )}
+      {selectedModal || null}
     </>
   );
 };
